feat: add creditOffer

+++ sql/credit.sql
@@ -1,4 +1,4 @@
 create table if not exists
     credit (credit_id UUID PRIMARY KEY NOT NULL,
             credit_limit INT,
-            credit_interest_rate INT);
+            credit_interest_rate DOUBLE);
